Function to check disjoint YEN routes
A function for check the disjoint YEN routes has implemented. A create Calls function for PDPP has implemented.

+++ include/Data/Parameters.h
@@ -214,7 +214,7 @@ public:
      * @brief Function to get Squeezing index (beta) used in the simulation 
      * with Protection option enabled.
      */
-    unsigned int GetBeta() const;
+    double GetBeta() const;
   /**
      * @brief Function to set Squeezing index (beta) used in the simulation
      * with Protection option enabled.
